Updated logging suppression
Added the log4j2.properties file to copy over to the spark containers to suppress logging info. The docker files have been changed accordingly. Added a note to the readme for those that don't want to rebuild containers, they can just copy over the file to the spark container manually.